fix!: correct the peer dependency requirement of the AirBnB React config
This was meant to be the latest version, but the base and React configs appear
to  be out of sync and I mistakenly assumed they would be the same. As always,
assumptions are the root of all evil.

While Airbnb <= 19 did not support ESLint 8, this was not specified in the
project dependencies so conflict resolution would NOT prevent us from being
installed alongside Airbnb 15. Subsequently, this is being marked as a breaking
change since it could *technically* cause a working downstream project to cross
major version boundaries.

Closes #19